Cleanup and documentation of the 2D ND space
